+++ examples/full-app-gourmet/views/views.go
@@ -25,7 +25,7 @@ func (rs Ressource) Routes(s *fuego.Server) {
 	// Public Chunks
 	fuego.Get(s, "/recipes-list", rs.showRecipesList)
 	fuego.Get(s, "/search", rs.searchRecipes)
-	fuego.Get(s, "/ingredients/preselect-unit", rs.unitPreselected).WithQueryParam("id", "")
+	fuego.Get(s, "/ingredients/preselect-unit", rs.unitPreselected).QueryParam("id", "")
 
 	// Admin Pages
 	adminRoutes := fuego.Group(s, "/admin")
