+++ go.work
@@ -6,4 +6,5 @@ use (
 	./examples/basic
 	./examples/full-app-gourmet
 	./examples/hello-world
+	./examples/openapi
 )
